iOS, 424 - Implement comprehensive TLS security for socket communication (#19)
* feat: Add SecurityOptions struct for TLS configuration

- Add SecurityOptions with useTLS, tokenProvider, and checksumProvider
- Make struct Sendable with @Sendable closures for actor safety
- Enable flexible security configuration with defaults to secure (TLS enabled)
- Support authentication tokens and message integrity verification

* feat: Add SecureMessageWrapper for message authentication and integrity

- Add SecureMessageWrapper to wrap CPRMessageModel with security metadata
- Include optional authentication tokens and checksums
- Add timestamp for message tracking
- Provide fluent API for adding tokens and checksums

* feat: Add TLS support and security configuration to TCPClient

- Add SecurityOptions parameter to TCPClient initializer
- Configure NWParameters based on useTLS setting (TLS vs TCP)
- Add createSecureMessage method for authentication and integrity
- Change from @Observable class to actor for thread safety
- Default to TLS enabled for secure connections

* refactor: Streamline message sending API with TLS integration

- Add sendMessage(_: CPRMessageModel) as primary type-safe method
- Add sendRawData(_: Data) for non-CPRMessageModel data
- Remove deprecated sendMessage(data: Data) to eliminate confusion
- Automatically apply TLS security features based on configuration
- Add smart security detection (TLS + tokens/checksums vs plain TLS)

* feat: Add TLS-aware connection establishment methods

- Add startConnection(with: CPRMessageModel) for type-safe initialization
- Add startConnection(value: Data) for raw data initialization
- Add startConnection() for connection without initial data
- Log connection type (TLS vs TCP) for debugging
- Integrate with security configuration automatically

* feat: Update connection ready handlers for TLS and typed messages

- Add connectionReady(message: CPRMessageModel) for typed messages
- Add connectionReady(data: Data) for raw data
- Add connectionReady() for no initial data scenarios
- Automatically use appropriate sending method (sendMessage vs sendRawData)

* feat: Add TLS support and security configuration to TCPServer

- Add SecurityOptions parameter to TCPServer initializer
- Configure NWParameters for TLS or TCP based on security settings
- Add createSecureMessage method for server-side security
- Add sendSecureMessage and broadcastSecureMessage methods
- Default to TLS enabled for secure server connections

* feat: Resolve actor isolation and add secure messaging support

- Make callback properties nonisolated(unsafe) for external assignment
- Add async initializer with proper logging
- Add updateLatestMessage method for message tracking
- Add createSecureMessage and sendSecureMessage methods
- Ensure thread safety while allowing callback configuration

* feat: Add secure message processing and smart routing

- Add sendMessage(_: CPRMessageModel, security: SecurityOptions) method
- Update processReceivedData to handle both secure and legacy messages
- Add automatic fallback from secure to regular message decoding
- Remove duplicate secure messaging methods (moved to main Connection class)
- Use await for proper async method calls

* refactor updateLatestMessage

* Update server and client tests for TLS and security features

* feat: change access control

* feat: create client tests

* feat: create connection tests

* feat: create server tests

* feat: implement generic TCP client/server with type-safe message handling

* refactor: redo tests to handle generic TCP client/server

* feat: fix merge errors

* fix: address swiftlint warnings

+++ Sources/RoboKit/Socket/JSON/Models/MessageModel.swift
@@ -15,13 +15,14 @@
 
 /// Message model containing claw control, position, and rotation data.
 public struct CPRMessageModel: Codable, Sendable, Equatable {
-    public init(clawControl: Bool, positionAndRotation: [Float], objectWidth: Float) {
+
+    public init(clawControl: Bool, positionAndRotation: [Double], objectWidth: Float) {
         self.clawControl = clawControl
         self.positionAndRotation = positionAndRotation
         self.objectWidth = objectWidth
     }
 
     public let clawControl: Bool
-    public let positionAndRotation: [Float]
+    public let positionAndRotation: [Double]
     public let objectWidth: Float
 }
